requirements.txt
Flask>=3.0.0
requests
python-dotenv
cryptography
Werkzeug
Jinja2
pywebview[qt]
yubico-client>=1.13.0
